requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
boto3==1.35.44
botocore==1.35.44
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi[standard]==0.115.2
fastapi-cli[standard]==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
passlib[bcrypt]==1.7.4
pika==1.3.2
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pyjwt==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
pyyaml==6.0.2
rich==13.9.2
s3transfer==0.10.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.40.0
typer==0.12.5
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn[standard]==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
